Validação de token colocado
Todas as Controllers agora tem as validações com token
Controller de Usuarios testada
Falta Clientes, animais e consultas

+++ app/Http/Controllers/Api/AnimaisController.php
@@ -14,6 +14,7 @@ class AnimaisController extends Controller
         $this->animal = $animal;
     }
 
+
     public function index()
     {
         // $data = ['data' => $this->animal->paginate(10)];
@@ -21,6 +22,16 @@ class AnimaisController extends Controller
         // $data = ['data' => "Coloque na URL o id do animal! Ex.: /api/animais/25"];
         $data = ['data' => 'Você precisa estar logado!'];
         return response()->json($data);
+    }
+
+    public function valida_token($token)
+    {
+        $result = TokensController::validar_token($token);
+        if (!$result) {
+            return false;
+            // return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
+        }
+        return true;
     }
 
     public function show(Animais $id)
@@ -47,16 +58,21 @@ class AnimaisController extends Controller
         */
         try {
             $animalData = $request->all();
-            if (isset($animalData['cliente_id'])) {
-                $animais_encontrados = Animais::where('cliente_id', request('cliente_id'))
-                    ->orderBy('nome_animal')
-                    ->get();
-                if (!$animais_encontrados->isEmpty()) {
-                    return response()->json(['data' => $animais_encontrados], 200);
+            if (isset($animalData['token'])) {
+                if ($this->valida_token(request('token'))) {
+                    if (isset($animalData['cliente_id'])) {
+                        $animais_encontrados = Animais::where('cliente_id', request('cliente_id'))
+                            ->orderBy('nome_animal')
+                            ->get();
+                        if (!$animais_encontrados->isEmpty()) {
+                            return response()->json(['data' => $animais_encontrados], 200);
+                        }
+                        return response()->json(['data' => 'Nenhum animal encontrado'], 200);
+                    }
+                    return response()->json(['data' => ['msg' => 'Falta um id do cliente']], 422);
                 }
-                return response()->json(['data' => 'Nenhum animal encontrado'], 200);
             }
-            return response()->json(['data' => ['msg' => 'Falta um id do cliente']], 422);
+            return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
         } catch (\Exception $e) {
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010), 500);
@@ -69,104 +85,109 @@ class AnimaisController extends Controller
     {
         try {
             $animalData = $request->all();
-            if (isset($animalData['cliente_id'])) {
-                //construindo a query
-                $query = Animais::query();
-                $query->where('cliente_id', request('cliente_id'));
-                $query->when(
-                    isset($animalData['nome_animal']),
-                    function ($q) {
-                        $nome_animal = request('nome_animal');
-                        return $q->where('nome_animal', 'like', "%$nome_animal%");
-                    }
-                );
-                $query->when(
-                    isset($animalData['dt_nasc']),
-                    function ($q) {
-                        $dt_nasc = request('dt_nasc');
-                        return $q->whereDate('dt_nasc', $dt_nasc);
-                    }
-                );
-                $query->when(
-                    isset($animalData['observacao']),
-                    function ($q) {
-                        $observacao = request('observacao');
-                        return $q->where('observacao', 'like', "%$observacao%");
-                    }
-                );
-                $query->when(
-                    isset($animalData['microchip']),
-                    function ($q) {
-                        $microchip = request('microchip');
-                        return $q->where('microchip', 'like', "%$microchip%");
-                    }
-                );
-                $query->when(
-                    isset($animalData['tag']),
-                    function ($q) {
-                        $tag = request('tag');
-                        return $q->where('tag', 'like', "%$tag%");
-                    }
-                );
-                $query->when(
-                    isset($animalData['sexo']),
-                    function ($q) {
-                        $sexo = request('sexo');
-                        return $q->where('sexo', 'like', $sexo);
-                    }
-                );
-                $query->when(
-                    isset($animalData['castrado']),
-                    function ($q) {
-                        $castrado = request('castrado');
-                        return $q->where('castrado', $castrado);
-                    }
-                );
-                $query->when(
-                    isset($animalData['cor']),
-                    function ($q) {
-                        $cor = request('cor');
-                        return $q->where('cor', $cor);
-                    }
-                );
-                $query->when(
-                    isset($animalData['created_at']),
-                    function ($q) {
-                        $created_at = request('created_at');
-                        return $q->whereDate('created_at',  $created_at);
-                    }
-                );
-                $query->when(
-                    isset($animalData['updated_at']),
-                    function ($q) {
-                        $updated_at = request('updated_at');
-                        return $q->whereDate('updated_at',  $updated_at);
-                    }
-                );
-                $query->orderBy('nome_animal');
-                $animais_encontrados = $query->get();
+            if (isset($animalData['token'])) {
+                if ($this->valida_token(request('token'))) {
+                    if (isset($animalData['cliente_id'])) {
+                        //construindo a query
+                        $query = Animais::query();
+                        $query->where('cliente_id', request('cliente_id'));
+                        $query->when(
+                            isset($animalData['nome_animal']),
+                            function ($q) {
+                                $nome_animal = request('nome_animal');
+                                return $q->where('nome_animal', 'like', "%$nome_animal%");
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['dt_nasc']),
+                            function ($q) {
+                                $dt_nasc = request('dt_nasc');
+                                return $q->whereDate('dt_nasc', $dt_nasc);
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['observacao']),
+                            function ($q) {
+                                $observacao = request('observacao');
+                                return $q->where('observacao', 'like', "%$observacao%");
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['microchip']),
+                            function ($q) {
+                                $microchip = request('microchip');
+                                return $q->where('microchip', 'like', "%$microchip%");
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['tag']),
+                            function ($q) {
+                                $tag = request('tag');
+                                return $q->where('tag', 'like', "%$tag%");
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['sexo']),
+                            function ($q) {
+                                $sexo = request('sexo');
+                                return $q->where('sexo', 'like', $sexo);
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['castrado']),
+                            function ($q) {
+                                $castrado = request('castrado');
+                                return $q->where('castrado', $castrado);
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['cor']),
+                            function ($q) {
+                                $cor = request('cor');
+                                return $q->where('cor', $cor);
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['created_at']),
+                            function ($q) {
+                                $created_at = request('created_at');
+                                return $q->whereDate('created_at',  $created_at);
+                            }
+                        );
+                        $query->when(
+                            isset($animalData['updated_at']),
+                            function ($q) {
+                                $updated_at = request('updated_at');
+                                return $q->whereDate('updated_at',  $updated_at);
+                            }
+                        );
+                        $query->orderBy('nome_animal');
+                        $animais_encontrados = $query->get();
 
-                /*
+                        /*
                 AQUI EM BAIXO NESSE FOREACH ERA UMA TENTATIVA DE 
                 LER O REQUEST E COLOCAR CADA PARAMETRO AUTOMATICAMENTE
                 NA QUERY, MAS DECIDI FAZER CADA UMA SEPARADA PARA MAPEAR CERTO
                 QUAIS OS PARAMETROS E OS CAMPOS DA CONSULTA, EVITANDO ASSIM BUGS.
                 */
-                // foreach ($animalData as $key => $value) {
-                //     $query->when(
-                //         isset($animalData[$key]),
-                //         function ($q, $value, $key) {
-                //             return $q->where($key, 'like', "%$value%");
-                //         }
-                //     );
-                // }
+                        // foreach ($animalData as $key => $value) {
+                        //     $query->when(
+                        //         isset($animalData[$key]),
+                        //         function ($q, $value, $key) {
+                        //             return $q->where($key, 'like', "%$value%");
+                        //         }
+                        //     );
+                        // }
 
-                if ($animais_encontrados->isEmpty()) {
-                    return response()->json(['data' => ["msg" => 'Nenhum animal encontrado']], 404);
+                        if ($animais_encontrados->isEmpty()) {
+                            return response()->json(['data' => ["msg" => 'Nenhum animal encontrado']], 404);
+                        }
+                        return response()->json(['data' => $animais_encontrados], 200);
+                    }
+                    return \response()->json(["data" => ['msg' => "Falta um id do cliente!"]], 422);
                 }
-                return response()->json(['data' => $animais_encontrados], 200);
             }
-            return \response()->json(["data" => ['msg' => "Falta um id do cliente!"]], 422);
+            return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
         } catch (\Exception $e) {
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010), 500);
@@ -179,8 +200,13 @@ class AnimaisController extends Controller
     {
         try {
             $animalData = $request->all();
-            $this->animal->create($animalData);
-            return response()->json(['data' => ['msg' => "Animal criado com sucesso"]], 201);
+            if (isset($animalData['token'])) {
+                if ($this->valida_token(request('token'))) {
+                    $this->animal->create($animalData);
+                    return response()->json(['data' => ['msg' => "Animal criado com sucesso"]], 201);
+                }
+            }
+            return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
         } catch (\Exception $e) {
             if ($e->getCode() == 'HY000') {
                 return response()->json(["data" => ["msg" => "Faltou o id do cliente ou o nome do animal", "code" => 1010]], 422);
@@ -202,12 +228,17 @@ class AnimaisController extends Controller
     {
         try {
             $animalData = $request->all();
-            $animal_encontrado = $this->animal->find($animalData['id']);
-            if (isset($animal_encontrado)) {
-                $animal_encontrado->update($animalData);
-                return response()->json(['data' => ['msg' => "Animal alterado com sucesso"]], 201);
+            if (isset($animalData['token'])) {
+                if ($this->valida_token(request('token'))) {
+                    $animal_encontrado = $this->animal->find($animalData['id']);
+                    if (isset($animal_encontrado)) {
+                        $animal_encontrado->update($animalData);
+                        return response()->json(['data' => ['msg' => "Animal alterado com sucesso"]], 201);
+                    }
+                    return response()->json(ApiError::errorMessage("Animal de id $animalData[id] nao encontrado", 404), 404);
+                }
             }
-            return response()->json(ApiError::errorMessage("Animal de id $animalData[id] nao encontrado", 404), 404);
+            return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
         } catch (\Exception $e) {
             if ($e->getCode() == 'HY000') {
                 return response()->json(["data" => ["msg" => "Faltou o id do cliente ou o nome do animal", "code" => 1010]], 422);
@@ -231,12 +262,17 @@ class AnimaisController extends Controller
         try {
             // dd($request->all());
             $animalData = $request->all();
-            $animal_encontrado = $this->animal->find(request('id'));
-            if (isset($animal_encontrado)) {
-                $animal_encontrado->delete($animalData);
-                return response()->json(['data' => ['msg' => "Animal " . $request['id'] . " deletado com sucesso"]], 200);
+            if (isset($animalData['token'])) {
+                if ($this->valida_token(request('token'))) {
+                    $animal_encontrado = $this->animal->find(request('id'));
+                    if (isset($animal_encontrado)) {
+                        $animal_encontrado->delete($animalData);
+                        return response()->json(['data' => ['msg' => "Animal " . $request['id'] . " deletado com sucesso"]], 200);
+                    }
+                    return response()->json(ApiError::errorMessage("Animal de id $animalData[id] nao encontrado", 404), 404);
+                }
             }
-            return response()->json(ApiError::errorMessage("Animal de id $animalData[id] nao encontrado", 404), 404);
+            return response()->json(["data" => ["msg" => ApiMessages::message(13)]], 422);
         } catch (\Exception $e) {
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010), 500);
